-refactored code for calculating transformations by consolidating all of the copy/pasted code in the computeTransformationBatch functions, and having the overloaded versions calling each other to end up at the same code. - simplified my own code for creating ImagePlus object from 2d arrays

+++ src/main/java/bunwarpj/BSplineModel.java
@@ -290,7 +290,7 @@ public class BSplineModel implements Runnable
 	 * @param isTarget enables the computation of the derivative or not
 	 */
 	public BSplineModel (
-			final int [][]img,
+			final double [][]img,
 			final boolean isTarget)
 	{
 		// Initialize thread
